Moved newcmd to shared
