--- backend_core/tomato/lib/cmd/bittorrent.py
@@ -1,1 +1,0 @@
-../../../../shared/lib/cmd/bittorrent.py
