+++ hostmanager/tomato/lib/newcmd/test/__init__.py
@@ -1,1 +1,1 @@
-__author__ = 'schwerdel'
+../../../../../shared/lib/newcmd/test/__init__.py
